Update
Fix issue in dashboard loading of navigation bar with name of user

fix the margin issue in preview page

made the navigation bar in preview page stick to top

Fixes with button padding and border color

Added title inputted by user to be the title for the page for dynamic urls

Change email to name in navigation bars

Added Profile icon in navigation bar

Some fixes in logic in change password

In login and signup page, fix the side image to be below the text

Upload image already works but  i have to use some generic ui first, i am having some logic errors here. The next action for this is to implement the UI for upload image

In Profile page, name and the email already appears in the dashboard but cannot be edited at the moment.